Move generate command into lib.
